Disk encryption frontend service updates
- Move disk encryption endpoint logic to dedicated service file
- Replace calling deprecated `PATCH`...`/mdm/apple/settings` for
    no-team disk encryption updates with call to `POST` ...`/fleet/disk_encryption`
    that now handles both team and no-team updates
