+++ frontend/services/entities/mdm.ts
@@ -1,5 +1,4 @@
 import {
-  DiskEncryptionStatus,
   IHostMdmProfile,
   IMdmCommandResult,
   IMdmProfile,
@@ -20,17 +19,6 @@ export interface IEulaMetadataResponse {
 }
 
 export type ProfileStatusSummaryResponse = Record<MdmProfileStatus, number>;
-
-export interface IDiskEncryptionStatusAggregate {
-  macos: number;
-  windows: number;
-  linux: number;
-}
-
-export type IDiskEncryptionSummaryResponse = Record<
-  DiskEncryptionStatus,
-  IDiskEncryptionStatusAggregate
->;
 
 export interface IGetProfilesApiParams {
   page?: number;
@@ -178,35 +166,6 @@ const mdmService = {
     }
 
     return sendRequest("GET", path);
-  },
-
-  getDiskEncryptionSummary: (teamId?: number) => {
-    let { MDM_DISK_ENCRYPTION_SUMMARY: path } = endpoints;
-
-    if (teamId) {
-      path = `${path}?${buildQueryStringFromParams({ team_id: teamId })}`;
-    }
-    return sendRequest("GET", path);
-  },
-
-  // TODO: API INTEGRATION: change when API is implemented that works for windows
-  // disk encryption too.
-  updateAppleMdmSettings: (enableDiskEncryption: boolean, teamId?: number) => {
-    const {
-      UPDATE_DISK_ENCRYPTION: teamsEndpoint,
-      CONFIG: noTeamsEndpoint,
-    } = endpoints;
-    if (teamId === 0) {
-      return sendRequest("PATCH", noTeamsEndpoint, {
-        mdm: {
-          enable_disk_encryption: enableDiskEncryption,
-        },
-      });
-    }
-    return sendRequest("POST", teamsEndpoint, {
-      enable_disk_encryption: enableDiskEncryption,
-      team_id: teamId,
-    });
   },
 
   initiateMDMAppleSSO: () => {
@@ -360,13 +319,6 @@ const mdmService = {
       DEVICE_USER_MDM_ENROLLMENT_PROFILE(token),
       undefined,
       "blob"
-    );
-  },
-  triggerLinuxDiskEncryptionKeyEscrow: (token: string) => {
-    const { DEVICE_TRIGGER_LINUX_DISK_ENCRYPTION_KEY_ESCROW } = endpoints;
-    return sendRequest(
-      "POST",
-      DEVICE_TRIGGER_LINUX_DISK_ENCRYPTION_KEY_ESCROW(token)
     );
   },
